Relocated all pdf related methods to seperate service
